refactor(Web): CPF/CNPJ document validation in production envs
